requirements.txt
apturl==0.5.2
attrs==21.4.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.0.4
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.0.3
future==0.18.2
httplib2==0.14.0
idna==2.8
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mock==4.0.3
monotonic==1.5
mysql-connector-python==8.0.28
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pluggy==1.0.0
protobuf==3.6.1
py==1.11.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyRFC3339==1.1
pytest==7.1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
testing.common.database==2.0.3
testing.mysqld==1.4.0
tomli==2.0.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==2.0.3
xkit==0.0.0
